+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.1.1] - 2017-06-18
 ### Changed
 - Allow patch version difference between local and remote (warn if not exactly the same)
 
@@ -48,7 +50,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Wait between last error log and exit, allows log to flush
 - Allow release builds to log debug lines (needed for verbose mode)
 
-[Unreleased]: https://github.com/daveallie/bindrs/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/daveallie/bindrs/compare/v0.1.1...HEAD
+[0.1.1]: https://github.com/daveallie/bindrs/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/daveallie/bindrs/compare/v0.0.4...v0.1.0
 [0.0.4]: https://github.com/daveallie/bindrs/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/daveallie/bindrs/compare/v0.0.2...v0.0.3